Añadir vista game
