+++ .babelrc
@@ -1,0 +1,5 @@
+{
+    "plugins": [
+        ["module:react-native-dotenv"]
+    ]
+}
