+++ .env
@@ -1,0 +1,1 @@
+TEST_MODE=UI
