+++ package.json
@@ -14,6 +14,7 @@
     "react": "17.0.2",
     "react-native": "0.66.4",
     "react-native-animatable": "^1.3.3",
+    "react-native-dotenv": "^3.3.1",
     "react-native-gesture-handler": "^2.2.0",
     "react-native-linear-gradient": "^2.5.6",
     "react-native-reanimated": "^2.4.1",
@@ -28,6 +29,7 @@
     "@trivago/prettier-plugin-sort-imports": "^3.1.1",
     "@types/jest": "^26.0.23",
     "@types/react-native": "^0.66.4",
+    "@types/react-native-dotenv": "^0.2.0",
     "@types/react-test-renderer": "^17.0.1",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
