+++ src/shared/defaultGameState.ts
@@ -1,0 +1,311 @@
+import ColorSchema from '../../assets/ColorSchema';
+import {PieceType} from './types';
+import {secondsSinceEpoch} from './utils';
+
+export const defaultPlayers = [
+  {
+    id: 0,
+    name: 'Player 1',
+    color: ColorSchema.yellow.normal,
+    score: 0,
+    hand: [
+      {
+        type: PieceType.KING,
+        quantity: 0,
+        isActive: false,
+      },
+      {
+        type: PieceType.QUEEN,
+        quantity: 45,
+        isActive: false,
+      },
+      {
+        type: PieceType.BISHOP,
+        quantity: 66,
+        isActive: false,
+      },
+      {
+        type: PieceType.KNIGTH,
+        quantity: 234,
+        isActive: false,
+      },
+      {
+        type: PieceType.ROOK,
+        quantity: 34,
+        isActive: false,
+      },
+      {
+        type: PieceType.PAWN,
+        quantity: 1,
+        isActive: true,
+      },
+    ],
+  },
+  {
+    id: 1,
+    name: 'Player 2',
+    color: ColorSchema.red.normal,
+    score: 0,
+    hand: [
+      {
+        type: PieceType.KING,
+        quantity: 0,
+        isActive: false,
+      },
+      {
+        type: PieceType.QUEEN,
+        quantity: 45,
+        isActive: false,
+      },
+      {
+        type: PieceType.BISHOP,
+        quantity: 66,
+        isActive: false,
+      },
+      {
+        type: PieceType.KNIGTH,
+        quantity: 234,
+        isActive: false,
+      },
+      {
+        type: PieceType.ROOK,
+        quantity: 34,
+        isActive: false,
+      },
+      {
+        type: PieceType.PAWN,
+        quantity: 1,
+        isActive: true,
+      },
+    ],
+  },
+  {
+    id: 2,
+    name: 'Player 3',
+    color: ColorSchema.blue.normal,
+    score: 0,
+    hand: [
+      {
+        type: PieceType.KING,
+        quantity: 0,
+        isActive: false,
+      },
+      {
+        type: PieceType.QUEEN,
+        quantity: 45,
+        isActive: false,
+      },
+      {
+        type: PieceType.BISHOP,
+        quantity: 66,
+        isActive: false,
+      },
+      {
+        type: PieceType.KNIGTH,
+        quantity: 234,
+        isActive: false,
+      },
+      {
+        type: PieceType.ROOK,
+        quantity: 34,
+        isActive: false,
+      },
+      {
+        type: PieceType.PAWN,
+        quantity: 1,
+        isActive: true,
+      },
+    ],
+  },
+  {
+    id: 3,
+    name: 'Player 4',
+    color: ColorSchema.green.normal,
+    score: 0,
+    hand: [
+      {
+        type: PieceType.KING,
+        quantity: 0,
+        isActive: false,
+      },
+      {
+        type: PieceType.QUEEN,
+        quantity: 45,
+        isActive: false,
+      },
+      {
+        type: PieceType.BISHOP,
+        quantity: 66,
+        isActive: false,
+      },
+      {
+        type: PieceType.KNIGTH,
+        quantity: 234,
+        isActive: false,
+      },
+      {
+        type: PieceType.ROOK,
+        quantity: 34,
+        isActive: false,
+      },
+      {
+        type: PieceType.PAWN,
+        quantity: 1,
+        isActive: true,
+      },
+    ],
+  },
+];
+
+export const defaultState = {
+  options: ['x2', 'pieza', '+100', '-100', 'efecto'],
+  messages: ['Hola xd', 'Que pro B)', 'Que noob'],
+  beginTime: secondsSinceEpoch(),
+  board: {
+    arrows: [{angle: 0}, {angle: 0}, {angle: 0}, {angle: 0}],
+    balls: [
+      [
+        {type: 1, position: {x: 0, y: 0}},
+        {type: 1, position: {x: 1, y: 1}},
+      ],
+      [
+        {type: 1, position: {x: 0, y: 5}},
+        {type: 1, position: {x: 1, y: 6}},
+      ],
+      [
+        {type: 1, position: {x: 5, y: 0}},
+        {type: 1, position: {x: 6, y: 1}},
+      ],
+      [
+        {type: 1, position: {x: 5, y: 5}},
+        {type: 1, position: {x: 6, y: 6}},
+      ],
+    ],
+    matrix: {
+      rows: 10,
+      cols: 10,
+      cells: [
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 10},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 0},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 10},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 0},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 10},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 0},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 10},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+        ],
+        [
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 100},
+          {health: 10},
+        ],
+      ],
+    },
+  },
+};
